feat: make distance sale

+++ database/migrations/2023_09_02_152618_create_stores_table.php
@@ -19,8 +19,8 @@ class CreateStoresTable extends Migration
             $table->string("email");
             $table->unsignedBigInteger("region_id");
             $table->unsignedBigInteger("manager_id");
-            $table->integer("lat");
-            $table->integer("lng");
+            $table->double("lat");
+            $table->double("lng");
             $table->timestamps();
             $table->softDeletes();
         });
